# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DB Auditor API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/analytics/findings-trends": {
      "get": {
        "description": "Contagens atuais exatas e série de achados observados na última execução analisada de cada período. Pontos parciais são identificados; períodos sem análise não viram zero.",
        "parameters": [
          {
            "name": "environment_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hour",
                "day",
                "week",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
...
    "/api/v1/analytics/kpis": {
      "get": {
        "description": "Contagens agregadas sem limite de lista. Cada ambiente usa uma execução de inventário coerente; inventory_status e notes indicam ausência ou coleta parcial.",
        "parameters": [
          {
            "name": "environment_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/api/v1/analytics/storage": {
      "get": {
        "description": "Distribuição atual e série de armazenamento da última execução por ambiente e período UTC. from/to aceitam RFC3339; janela máxima de 366 dias, ou 7 dias para hora.",
        "parameters": [
          {
            "name": "environment_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hour",
                "day",
                "week",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
...
    "/api/v1/environments/{id}/capabilities": {
      "get": {
        "description": "Matriz de capacidades observadas para o mecanismo e a execução mais recente.",
        "responses": {
          "200": {
            "description": "Resultado"
          },
          "400": {
            "description": "Entrada inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "422": {
            "description": "Escopo ou comparação indisponível"
          }
        }
      }
    },
...
  },
...
}